type fix

+++ tests/test_merge.py
@@ -13,19 +13,19 @@ class TestMerge(unittest.TestCase):
         self.grid0 = np.array([[2, 2, 2, 2],
                                [4, 4, 0, 0],
                                [2, 4, 8, 16],
-                               [2, 0, 2, 4]], dtype=np.float64)
+                               [2, 0, 2, 4]])
 
         self.grid1 = np.array([[2, 2, 4, 2],
                                [8, 4, 4, 16],
                                [8, 0, 0, 16],
-                               [0, 0, 0, 32]], dtype=np.float64)
+                               [0, 0, 0, 32]])
 
     def test_merge_left_grid0(self):
         grid, _ = merge_left(self.grid0)
         grid_res = np.array([[4, 4, 0, 0],
                              [8, 0, 0, 0],
                              [2, 4, 8, 16],
-                             [4, 4, 0, 0]], dtype=np.float64)
+                             [4, 4, 0, 0]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_left_grid1(self):
@@ -33,7 +33,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[4, 4, 2, 0],
                              [8, 8, 16, 0],
                              [8, 16, 0, 0],
-                             [32, 0, 0, 0]], dtype=np.float64)
+                             [32, 0, 0, 0]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_right_grid0(self):
@@ -41,7 +41,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[0, 0, 4, 4],
                              [0, 0, 0, 8],
                              [2, 4, 8, 16],
-                             [0, 0, 4, 4]], dtype=np.float64)
+                             [0, 0, 4, 4]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_right_grid1(self):
@@ -49,7 +49,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[0, 4, 4, 2],
                              [0, 8, 8, 16],
                              [0, 0, 8, 16],
-                             [0, 0, 0, 32]], dtype=np.float64)
+                             [0, 0, 0, 32]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_up_grid0(self):
@@ -57,7 +57,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[2, 2, 2, 2],
                              [4, 8, 8, 16],
                              [4, 0, 2, 4],
-                             [0, 0, 0, 0]], dtype=np.float64)
+                             [0, 0, 0, 0]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_up_grid1(self): 
@@ -65,7 +65,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[2, 2, 8, 2],
                              [16, 4, 0, 32],
                              [0, 0, 0, 32],
-                             [0, 0, 0, 0]], dtype=np.float64)
+                             [0, 0, 0, 0]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_down_grid0(self):
@@ -73,7 +73,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[0, 0, 0, 0],
                              [2, 0, 2, 2],
                              [4, 2, 8, 16],
-                             [4, 8, 2, 4]], dtype=np.float64)
+                             [4, 8, 2, 4]])
         self.assertTrue(np.all(grid == grid_res))
 
     def test_merge_down_grid1(self):
@@ -81,7 +81,7 @@ class TestMerge(unittest.TestCase):
         grid_res = np.array([[0, 0, 0, 0],
                              [0, 0, 0, 2],
                              [2, 2, 0, 32],
-                             [16, 4, 8, 32]], dtype=np.float64)
+                             [16, 4, 8, 32]])
         self.assertTrue(np.all(grid == grid_res))
